Add files via upload
Welcome to the Financial Time Series Project branch! This branch contains all the essential code and documentation related to the project. Here, you will find:

Code: Scripts and modules developed for analyzing and forecasting financial time series data.
Documentation: Detailed guides and explanations on the methodologies used, as well as instructions on how to set up, run, and contribute to the project.
Feel free to explore the contents, provide feedback, and contribute to the development and improvement of the project. Your collaboration is highly appreciated!